refacto into proper components folder
